refactor: merge role and session auth context
