+++ src/features/auth/ProtectedRoute.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useAuth } from './AuthContext';
+import { useAuth } from '../../contexts/AuthContext';
 import { Navigate, useLocation } from 'react-router-dom';
 
 export default function ProtectedRoute({ children, roles }) {
